requirements.txt
python-telegram-bot==20.7
web3==6.11.5
